requirements.txt
FastAPI>=0.103.0,<0.110.0
uvicorn[standard]>=0.23.2,<0.28.0
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
SQLAlchemy>=2.0.20,<2.1.0
alembic>=1.12.0,<2.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
PyJWT>=2.8.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
pytest>=7.4.2,<8.0.0
httpx>=0.25.0,<0.28.0
email-validator>=2.0.0,<3.0.0
bcrypt<4.0.0
